Lots of stuff. Can't write commit message, have to do more slave labour. Also so much stuff that I actually cant remember all of it. Biggest addition is probably the grenade system tho.
Grenades blow stuff up. Good fun.

Kill me.

+++ ConsoleDungeonCrawler/ConsoleDungeonCrawler/Vector2.cs
@@ -13,7 +13,7 @@ public class Vector2
         this.y = 0;
     }
 
-    public Vector2(int x, int y)
+    public Vector2(float x, float y)
     {
         this.x = x;
         this.y = y;
